Adding a pin in accountInfo and adding a password for the user

+++ src/main/java/com/revature/CLIBank/model/Transaction.java
@@ -12,19 +12,15 @@ public class Transaction {
     private UUID sourceAccountId;
     private UUID destinationAccountId;
     private TransactionType type;
-    private double amount;
+    private long amount;
     private LocalDateTime timestamp;
 
     /**
      * Creates a new transfer between two accounts.
      * Generates a unique ID and sets the timestamp to now.
      *
-     * @param sourceAccountId account money leaves
-     * @param destinationAccountId account money enters
-     * @param type should be transfer
-     * @param amount amount moved
      */
-    public Transaction(UUID sourceAccountId, UUID destinationAccountId, TransactionType type, double amount) {
+    public Transaction(UUID sourceAccountId, UUID destinationAccountId, TransactionType type, long amount) {
         this.transactionId = UUID.randomUUID(); 
         this.sourceAccountId = sourceAccountId;
         this.destinationAccountId = destinationAccountId;
@@ -36,11 +32,8 @@ public class Transaction {
     /**
      * Creates a new deposit or withdrawal on a single account.
      * Generates a unique ID, sets destination and timestamps now.
-     * @param sourceAccountId account being deposited to or withdrawn from
-     * @param type Type of transaction
-     * @param amount amount moved
      */
-    public Transaction(UUID sourceAccountId, TransactionType type, double amount) {
+    public Transaction(UUID sourceAccountId, TransactionType type, long amount) {
         this.transactionId = UUID.randomUUID();
         this.sourceAccountId = sourceAccountId;
         this.destinationAccountId = null;
@@ -51,14 +44,8 @@ public class Transaction {
 
     /**
      * Rebuilds a transaction from an existing database row.
-     * @param transactionId the ID already stored in Transactions
-     * @param sourceAccountId account money left
-     * @param destinationAccountId account money entered
-     * @param type deposit, withdrawal, or transfer
-     * @param amount amount moved
-     * @param timestamp when the transaction occurred
      */
-    public Transaction(UUID transactionId, UUID sourceAccountId, UUID destinationAccountId, TransactionType type, double amount, LocalDateTime timestamp) {
+    public Transaction(UUID transactionId, UUID sourceAccountId, UUID destinationAccountId, TransactionType type, long amount, LocalDateTime timestamp) {
         this.transactionId = transactionId;
         this.sourceAccountId = sourceAccountId;
         this.destinationAccountId = destinationAccountId;
@@ -99,11 +86,11 @@ public class Transaction {
         this.type = type;
     }
 
-    public double getAmount() {
+    public long getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(long amount) {
         this.amount = amount;
     }
